requirements.txt
torch~=2.5.1
torchvision~=0.20.1
omegaconf==2.1.2
wandb~=0.19.5
pillow~=11.1.0
python-dotenv~=1.0.1